+++ app/src/main/java/com/likeminds/feedsx/posttypes/model/AttachmentType.kt
@@ -1,4 +1,4 @@
-package com.likeminds.feedsx.post.model
+package com.likeminds.feedsx.posttypes.model
 
 import androidx.annotation.IntDef
