+++ app/src/main/java/com/likeminds/feedsx/posttypes/util/PostTypeUtil.kt
@@ -1,4 +1,4 @@
-package com.likeminds.feedsx.post.util
+package com.likeminds.feedsx.posttypes.util
 
 import android.text.*
 import android.text.style.ClickableSpan
@@ -13,13 +13,13 @@ import com.likeminds.feedsx.R
 import com.likeminds.feedsx.branding.model.BrandingData
 import com.likeminds.feedsx.databinding.*
 import com.likeminds.feedsx.overflowmenu.view.OverflowMenuPopup
-import com.likeminds.feedsx.post.adapter.DocumentsPostAdapter
-import com.likeminds.feedsx.post.adapter.MultipleMediaPostAdapter
-import com.likeminds.feedsx.post.adapter.PostAdapter.PostAdapterListener
-import com.likeminds.feedsx.post.model.AttachmentViewData
-import com.likeminds.feedsx.post.model.IMAGE
-import com.likeminds.feedsx.post.model.PostViewData
-import com.likeminds.feedsx.post.model.VIDEO
+import com.likeminds.feedsx.posttypes.view.adapter.DocumentsPostAdapter
+import com.likeminds.feedsx.posttypes.view.adapter.MultipleMediaPostAdapter
+import com.likeminds.feedsx.posttypes.view.adapter.PostAdapter.PostAdapterListener
+import com.likeminds.feedsx.posttypes.model.AttachmentViewData
+import com.likeminds.feedsx.posttypes.model.IMAGE
+import com.likeminds.feedsx.posttypes.model.PostViewData
+import com.likeminds.feedsx.posttypes.model.VIDEO
 import com.likeminds.feedsx.utils.MemberImageUtil
 import com.likeminds.feedsx.utils.TimeUtil
 import com.likeminds.feedsx.utils.ViewUtils
