+++ app/src/main/java/com/likeminds/feedsx/posttypes/view/adapter/DocumentsPostAdapter.kt
@@ -1,7 +1,7 @@
-package com.likeminds.feedsx.post.adapter
+package com.likeminds.feedsx.posttypes.view.adapter
 
-import com.likeminds.feedsx.post.adapter.PostAdapter.PostAdapterListener
-import com.likeminds.feedsx.post.adapter.databinder.postdocuments.ItemDocumentViewDataBinder
+import com.likeminds.feedsx.posttypes.view.adapter.PostAdapter.PostAdapterListener
+import com.likeminds.feedsx.posttypes.view.adapter.databinder.postdocuments.ItemDocumentViewDataBinder
 import com.likeminds.feedsx.utils.customview.BaseRecyclerAdapter
 import com.likeminds.feedsx.utils.customview.ViewDataBinder
 import com.likeminds.feedsx.utils.model.BaseViewType
