+++ app/src/main/java/com/likeminds/feedsx/posttypes/view/adapter/databinder/ItemPostSingleVideoViewDataBinder.kt
@@ -1,16 +1,16 @@
-package com.likeminds.feedsx.post.adapter.databinder
+package com.likeminds.feedsx.posttypes.view.adapter.databinder
 
 import android.media.MediaPlayer.OnPreparedListener
 import android.net.Uri
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.likeminds.feedsx.databinding.ItemPostSingleVideoBinding
-import com.likeminds.feedsx.overflowmenu.adapter.OverflowMenuAdapterListener
+import com.likeminds.feedsx.overflowmenu.view.adapter.OverflowMenuAdapterListener
 import com.likeminds.feedsx.overflowmenu.model.OverflowMenuItemViewData
 import com.likeminds.feedsx.overflowmenu.view.OverflowMenuPopup
-import com.likeminds.feedsx.post.adapter.PostAdapter.PostAdapterListener
-import com.likeminds.feedsx.post.model.PostViewData
-import com.likeminds.feedsx.post.util.PostTypeUtil
+import com.likeminds.feedsx.posttypes.view.adapter.PostAdapter.PostAdapterListener
+import com.likeminds.feedsx.posttypes.model.PostViewData
+import com.likeminds.feedsx.posttypes.util.PostTypeUtil
 import com.likeminds.feedsx.utils.ViewUtils.hide
 import com.likeminds.feedsx.utils.customview.ViewDataBinder
 import com.likeminds.feedsx.utils.model.ITEM_POST_SINGLE_VIDEO
