+++ app/src/main/java/com/likeminds/feedsx/posttypes/view/adapter/databinder/postdocuments/ItemDocumentViewDataBinder.kt
@@ -1,10 +1,10 @@
-package com.likeminds.feedsx.post.adapter.databinder.postdocuments
+package com.likeminds.feedsx.posttypes.view.adapter.databinder.postdocuments
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.likeminds.feedsx.databinding.ItemDocumentBinding
-import com.likeminds.feedsx.post.adapter.PostAdapter.PostAdapterListener
-import com.likeminds.feedsx.post.model.AttachmentViewData
+import com.likeminds.feedsx.posttypes.view.adapter.PostAdapter.PostAdapterListener
+import com.likeminds.feedsx.posttypes.model.AttachmentViewData
 import com.likeminds.feedsx.utils.customview.ViewDataBinder
 import com.likeminds.feedsx.utils.model.ITEM_POST_DOCUMENTS_ITEM
