+++ app/src/main/res/layout/activity_main.xml
@@ -2,7 +2,7 @@
 <layout xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     xmlns:android="http://schemas.android.com/apk/res/android"
-    tools:context=".MainActivity">
+    tools:context=".feed.view.MainActivity">
 
     <data>
         <import type="android.view.View" />
